+++ src/main/java/com/example/enterprise_internet_applications_project/repositories/GroupRepository.java
@@ -1,5 +1,6 @@
-package com.example.enterprise_internet_applications_project.group;
+package com.example.enterprise_internet_applications_project.repositories;
 
+import com.example.enterprise_internet_applications_project.models.Group;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
